Fix test failure
