+++ DragaliaAPI.Integration.Test/Dragalia/FortTest.cs
@@ -226,7 +226,7 @@ public class FortTest : TestFixture
         result.build_start_date.Should().NotBe(DateTimeOffset.UnixEpoch);
         result.build_end_date.Should().NotBe(DateTimeOffset.UnixEpoch);
         result.build_end_date.Should().BeAfter(result.build_start_date);
-        response.fort_detail.working_carpenter_num.Should().Be(1);
+        response.fort_detail.working_carpenter_num.Should().BeGreaterThan(0); // Requiring '1' can conflict with other building tests
     }
 
     [Fact]
